control of i2c devices through gui possible
also fixed some bugs with closing the gui application without disconnecting first. still no idea how to send a disconnect signal via tcp on closing the demon with strg c.

+++ shared/i2cproperty.cpp
@@ -1,7 +1,7 @@
 #include <i2cproperty.h>
 
 
-I2cProperty::I2cProperty(qint8 pcaChannel, float dac_Thresh1, float dac_Thresh2, float biasVoltage, bool biasPowerOn){
+I2cProperty::I2cProperty(int pcaChannel, float dac_Thresh1, float dac_Thresh2, float biasVoltage, bool biasPowerOn){
     pcaChann = pcaChannel;
     thresh1 = dac_Thresh1;
     thresh2 = dac_Thresh2;
