+++ shared/i2cproperty.h
@@ -2,8 +2,8 @@
 #define I2CPROPERTY_H
 #include <QDataStream>
 struct I2cProperty{
-    I2cProperty(qint8 pcaChannel = -1, float dac_Thresh1 = -1, float dac_Thresh2 = -1, float biasVoltage = -1, bool biasPowerOn = false);
-    quint8 pcaChann;
+    I2cProperty(int pcaChannel = -1, float dac_Thresh1 = -1, float dac_Thresh2 = -1, float biasVoltage = -1, bool biasPowerOn = false);
+    int pcaChann;
     float thresh1, thresh2, bias_Voltage;
     bool bias_powerOn;
 };
